fixed js
