+++ app/javascript/plugins/flatpickr.js
@@ -1,9 +1,16 @@
 import flatpickr from "flatpickr";
 
-flatpickr(".datepicker", {
-  altInput: true,
-  altFormat: "F j, Y",
-  dateFormat: "Y-m-d",
-  minDate: "today",
-  // enableTime: true,
-});
+const initFlatPicker = () => {
+  const flatpickr = document.querySelector('.datepickr')
+    if (flatpickr) {
+      flatpickr(".datepicker", {
+        altInput: true,
+        altFormat: "F j, Y",
+        dateFormat: "Y-m-d",
+        minDate: "today",
+        // enableTime: true,
+      });
+    }
+}
+
+export { initFlatPicker }
